requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
PyYAML
faiss-cpu
openai==0.27.0
pandas
numpy

